+++ BattleNetwork/bnGraveyardBackground.cpp
@@ -19,17 +19,17 @@ GraveyardBackground::~GraveyardBackground(void) {
 }
 
 void GraveyardBackground::Update(float _elapsed) {
-  progress += 0.2 * _elapsed;
+  progress += 0.2f * _elapsed;
   if (progress >= 1.f) progress = 0.0f;
 
-  y += 0.5 * _elapsed;
-  x += 0.5 *_elapsed;
+  y += 0.5f * _elapsed;
+  x += 0.5f *_elapsed;
 
   if (x > 1) x = 0;
   if (y > 1) y = 0;
 
-  int frame = progress * COMPONENT_FRAME_COUNT;
+  int frame = (int)(progress * COMPONENT_FRAME_COUNT);
 
   Wrap(sf::Vector2f(x, y));
-  TextureOffset(sf::Vector2f(frame*COMPONENT_WIDTH, 0));
+  TextureOffset(sf::Vector2f((float)(frame*COMPONENT_WIDTH), 0));
 }
